Adds group room creation and user addition functionality
Introduces a new `GroupRoomService` to handle group room creation and user addition to existing rooms.
Migrates group room creation logic to the new service.
Removes redundant `RoomService` methods related to group rooms.
Creates necessary service implementation and repository classes for group rooms.
