+++ api/Chat/src/main/java/com/lemoo/chat/controller/GroupController.java
@@ -10,27 +10,33 @@ package com.lemoo.chat.controller;
 import com.lemoo.chat.dto.common.AuthenticatedAccount;
 import com.lemoo.chat.dto.request.GroupRoomRequest;
 import com.lemoo.chat.dto.response.ApiResponse;
-import com.lemoo.chat.service.RoomService;
+import com.lemoo.chat.service.GroupRoomService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/rooms")
 @RequiredArgsConstructor
 public class GroupController {
-    private final RoomService roomService;
+    private final GroupRoomService groupRoomService;
 
     @PostMapping
     public ApiResponse<Boolean> createRoom(
             @RequestBody @Valid GroupRoomRequest request,
             @AuthenticationPrincipal AuthenticatedAccount account
     ) {
-        return ApiResponse.success(roomService.createGroupRoom(request, account));
+        return ApiResponse.success(groupRoomService.createGroupRoom(request, account));
+    }
+
+    @PostMapping("{roomId}")
+    public ApiResponse<Boolean> addUserToGroup(
+            @RequestParam("invitee") String inviteeId,
+            @PathVariable("roomId") String roomId,
+            @AuthenticationPrincipal AuthenticatedAccount account
+    ) {
+        return ApiResponse.success(groupRoomService.addUserToRoom(inviteeId, roomId, account));
     }
 
 }
