+++ api/Chat/src/main/java/com/lemoo/chat/service/RoomService.java
@@ -7,7 +7,6 @@
 package com.lemoo.chat.service;
 
 import com.lemoo.chat.dto.common.AuthenticatedAccount;
-import com.lemoo.chat.dto.request.GroupRoomRequest;
 import com.lemoo.chat.dto.response.PageableResponse;
 import com.lemoo.chat.dto.response.RoomDetailResponse;
 import com.lemoo.chat.dto.response.RoomResponse;
@@ -22,6 +21,5 @@ public interface RoomService {
     RoomDetailResponse getRoomDetail(String roomId, AuthenticatedAccount account);
 
     Room findRoomById(String roomId);
-
-    boolean createGroupRoom(GroupRoomRequest request, AuthenticatedAccount account);
+    
 }
